feat: Merge CacheWarmingCard and WarmingQueueCard with bulk delete (#122)
Merge WarmingQueueCard into CacheWarmingCard:
- Queue table with pause/resume/delete actions
- Checkbox selection for bulk operations
- Bulk delete with confirmation modal
- Cancel button for running jobs
- Estimated time remaining display
- Queries per second metric

Update API endpoints to new pattern:
- /api/admin/cache/warm-* → /api/admin/warming/*
- Add bulkDeleteWarmingJobs(), cancelWarmingJob()
- SSE reconnection with lastEventId param

Update store:
- Add selectedJobIds for bulk selection
- Add lastEventId for SSE reconnection
- Persist lastEventId to localStorage

Remove WarmingQueueCard.tsx (merged into CacheWarmingCard)
Update SettingsView to use single unified component

Closes #122
