+++ frontend/src/components/features/admin/index.ts
@@ -14,4 +14,3 @@ export { CacheStatsCard } from './CacheStatsCard';
 export { CacheTopQueriesCard } from './CacheTopQueriesCard';
 export { CacheWarmingCard } from './CacheWarmingCard';
 export { ClearCacheModal } from './ClearCacheModal';
-export { WarmingQueueCard } from './WarmingQueueCard';
